## Summary
Updates the TokenPricing model to use float data types instead of
integers for input and output token pricing attributes, enabling more
precise cost calculations.

## Key Changes
- Modified TokenPricing model attribute types from `int` to `float`
- Enhanced precision for token cost calculations
- Improved accuracy in pricing computations

## Impact
- **Breaking Change**: This change modifies the data model and may
affect existing integrations that expect integer values
- Applications consuming the TokenPricing model will now receive float
values instead of integers
- Database schema may require migration if persisting these values

## Testing Considerations
- Verify that all token pricing calculations work correctly with float
values
- Test backward compatibility with existing API consumers
- Validate precision improvements in cost computation scenarios
- Ensure proper handling of decimal values in pricing workflows

## Technical Notes
- The change affects the core pricing model used throughout the
robosystems client
- Consider updating any validation rules that may have assumed integer
constraints
- Review serialization/deserialization logic to ensure proper float
handling

---

**Branch Info:**
- Source: `bugfix/token-float`
- Target: `main`
- Type: bugfix

+++ robosystems_client/models/token_pricing.py
@@ -14,12 +14,12 @@ class TokenPricing:
   """AI token pricing for a specific model.
 
   Attributes:
-      input_per_1k_tokens (int): Credits per 1K input tokens
-      output_per_1k_tokens (int): Credits per 1K output tokens
+      input_per_1k_tokens (float): Credits per 1K input tokens
+      output_per_1k_tokens (float): Credits per 1K output tokens
   """
 
-  input_per_1k_tokens: int
-  output_per_1k_tokens: int
+  input_per_1k_tokens: float
+  output_per_1k_tokens: float
   additional_properties: dict[str, Any] = _attrs_field(init=False, factory=dict)
 
   def to_dict(self) -> dict[str, Any]:
